+++ shared/src/commonMain/kotlin/co/touchlab/kampkit/base/StaleDataDelegate.kt
@@ -1,11 +1,23 @@
 package co.touchlab.kampkit.base
 
+import com.russhwolf.settings.Settings
+import kotlinx.datetime.Clock
+
 /**
- * Handles updating/checking if data is stale given a [StaleDataKey]
+ * Handles updating/checking if data is stale given a [StaleDataKey] based on system time
+ * and stored [settings].
  */
-interface StaleDataDelegate {
+class StaleDataDelegate(
+    private val settings: Settings,
+    private val clock: Clock
+) {
 
-    fun updateLastTime(key: StaleDataKey)
+    fun updateLastTime(key: StaleDataKey) {
+        settings.putLong(key.keyName, clock.now().toEpochMilliseconds())
+    }
 
-    fun isDataStale(key: StaleDataKey): Boolean
+    fun isDataStale(key: StaleDataKey): Boolean {
+        val lastDownloadTimeMS = settings.getLong(key.keyName, 0)
+        return lastDownloadTimeMS + key.msUntilStale < clock.now().toEpochMilliseconds()
+    }
 }
