+++ shared/src/commonMain/kotlin/co/touchlab/kampkit/feature/breed/BreedRepository.kt
@@ -1,15 +1,79 @@
 package co.touchlab.kampkit.feature.breed
 
+import co.touchlab.kampkit.DatabaseHelper
+import co.touchlab.kampkit.base.ApiStatus
+import co.touchlab.kampkit.base.StaleDataDelegate
+import co.touchlab.kampkit.base.StaleDataKey
 import co.touchlab.kampkit.db.Breed
+import co.touchlab.kampkit.ktor.Api
+import co.touchlab.kampkit.networkmodels.BreedDto
+import co.touchlab.kermit.Logger
+import co.touchlab.stately.ensureNeverFrozen
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.flow.combine
 
-interface BreedRepository {
+class BreedRepository(
+    private val log: Logger,
+    private val dbHelper: DatabaseHelper,
+    private val api: Api,
+    private val staleDataDelegate: StaleDataDelegate
+) {
 
-    fun getData(): Flow<BreedRepoData>
+    init {
+        ensureNeverFrozen()
+    }
 
-    suspend fun refreshDataIfStale()
+    private val requestStatus = MutableStateFlow<ApiStatus<BreedDto>>(ApiStatus.NoAction)
 
-    suspend fun refreshData()
+    fun getData(): Flow<BreedRepoData> =
+        combine(dbHelper.selectAllItems(), requestStatus.asStateFlow()) { items, requestStatus ->
+            when (requestStatus) {
+                is ApiStatus.Error -> {
+                    BreedRepoData(isError = true)
+                }
+                ApiStatus.NoAction,
+                ApiStatus.NetworkError,
+                is ApiStatus.Success -> {
+                    BreedRepoData(
+                        data = items.sortedBy { it.name }
+                    )
+                }
+                ApiStatus.Loading -> BreedRepoData(isLoading = true)
+            }
+        }
 
-    suspend fun updateBreedFavorite(breed: Breed)
+    suspend fun refreshDataIfStale() {
+        log.d { "Checking if data is stale." }
+        if (staleDataDelegate.isDataStale(StaleDataKey.BreedStaleData)) {
+            refreshData()
+        }
+    }
+
+    suspend fun refreshData() {
+        requestStatus.tryEmit(ApiStatus.Loading)
+
+        log.d { "Refreshing data." }
+
+        val result = api.getBreeds()
+        staleDataDelegate.updateLastTime(StaleDataKey.BreedStaleData)
+
+        val data = result as? ApiStatus.Success
+        data?.let {
+            persistData(it.data)
+        }
+
+        log.d { "Network result: $result" }
+
+        requestStatus.tryEmit(result)
+    }
+
+    private suspend fun persistData(data: BreedDto) {
+        dbHelper.insertBreeds(data.message.keys.toList())
+    }
+
+    suspend fun updateBreedFavorite(breed: Breed) {
+        dbHelper.updateFavorite(breed.id, !breed.favorite)
+    }
 }
